Update usage from the `node` binary

+++ executor/limiter.go
@@ -3,6 +3,6 @@ package executor
 // noopLimiter is a dummy limiter used when processes run without any resource limitations.
 type noopLimiter struct{}
 
-func (n *noopLimiter) LimitProcess(pid int) error {
+func (n *noopLimiter) LimitProcess(pid uint64) error {
 	return nil
 }
