+++ executor/limits.go
@@ -1,5 +1,5 @@
 package executor
 
 type Limiter interface {
-	LimitProcess(pid int) error
+	LimitProcess(pid uint64) error
 }
